tcnet/requirements.txt
certifi==2018.8.24
chardet==3.0.4
clearbit==0.1.7
Django==2.0
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
Markdown==3.0.1
pep8==1.7.1
pyhunter==0.3
PyJWT==1.6.4
pytz==2018.5
requests==2.19.1
urllib3==1.23
